The Doctrine of the Fourth Commandement
The doctrine of the Fourth Commandement, deformed by popery, reformed & restored to its primitive purity wherein is clearely proved by Scripture, arguments, and reasons, that the seventh day of the week, and not the first, viz. the day called Saturday (and not the day called Sunday) is the true Christian Sabbath ... / objections answered, and the truth cleared, by Gods unworthy servant, J.O. Author: Ockford, James. 1650